fix(appointments): resolve GraphQL 400 error by standardizing `id` field in schema and queries
