+++ frontend/src/graphql/queries/patientQueries.ts
@@ -1,0 +1,16 @@
+import { gql } from '@apollo/client';
+
+export const GET_PATIENT_APPOINTMENTS = gql`
+  query GetAppointmentsForPatient {
+    getAppointmentsForPatient {
+      id
+      doctor {
+        email
+      }
+      date
+      time
+      status
+      notes
+    }
+  }
+`;
