--- frontend/src/pages/DoctorAppointments.tsx
@@ -1,3 +1,0 @@
-export default function DoctorAppointments() {
-  return <div className="p-8">📅 Doctor Appointments Page</div>;
-}
