+++ frontend/src/pages/doctor-appointment/index.tsx
@@ -1,0 +1,4 @@
+export default function DoctorAppointments() {
+    return <div className="p-8">📅 Doctor Appointments Page</div>;
+  }
+
