+++ frontend/src/router/Router.tsx
@@ -3,13 +3,13 @@ import { useAuth } from '@/contexts/useAuth';
 
 import Signup from '@/pages/Signup';
 import Login from '@/pages/login';
-import ForgotPassword from '@/pages/ForgotPasswordPage';
-import ResetPassword from '@/pages/ResetPasswordPage';
-import Profile from '@/pages/Profile';
-import DoctorDashboard from '@/pages/DoctorDashboard';
-import PatientDashboard from '@/pages/PatientDashboard';
-import DoctorAppointments from '@/pages/DoctorAppointments';
-import PatientAppointments from '@/pages/PatientAppointments';
+import ForgotPassword from '@/pages/forgot-password';
+import ResetPassword from '@/pages/reset-password';
+import Profile from '@/pages/dashboard/profile';
+import DoctorDashboard from '@/pages/dashboard/doctor';
+import PatientDashboard from '@/pages/dashboard/patient';
+import DoctorAppointments from '@/pages/doctor-appointment';
+import PatientAppointments from '@/pages/patient-appoimtment';
 import PrivateRoute from '@/components/auth/PrivateRoute';
 import RoleBasedRoute from '@/components/auth/RoleBasedRoute';
 import PageNotFound from '@/pages/PageNotFound';
